+++ version_history.md
@@ -1,3 +1,8 @@
+### **New on version 5.0**
+* Remove TimeOut with argument function and remove MicroTuple dependency
+* Remove print container
+
+
 ### **New on version 4.0**
 * change function pointer to template, enableling to use of functor and lambdas
 * Remove virtual TO_callbackTrigguer function